# internal/api/swagger/orders.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "orders_api/orders.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "OrdersService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/health": {
      "get": {
        "operationId": "OrdersService_Health",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1HealthResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "OrdersService"
        ]
      }
    },
    "/orders": {
      "post": {
        "summary": "Create / update / soft-delete order.",
        "operationId": "OrdersService_UpsertOrder",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpsertOrderResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1UpsertOrderRequest"
            }
          }
        ],
        "tags": [
          "OrdersService"
        ]
      }
    },
    "/orders/by-external/{externalId}": {
      "get": {
        "operationId": "OrdersService_GetOrderByExternalID",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetOrderByExternalIDResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "externalId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "userId",
            "description": "query param",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "OrdersService"
        ]
      }
    },
    "/orders/{orderId}": {
      "get": {
        "operationId": "OrdersService_GetOrderByID",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetOrderByIDResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "OrdersService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1GetOrderByExternalIDResponse": {
      "type": "object",
      "properties": {
        "order": {
          "$ref": "#/definitions/v1OrderModel"
        }
      }
    },
    "v1GetOrderByIDResponse": {
      "type": "object",
      "properties": {
        "order": {
          "$ref": "#/definitions/v1OrderModel"
        }
      }
    },
    "v1HealthResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "title": "\"ok\""
        }
      }
    },
    "v1OrderModel": {
      "type": "object",
      "properties": {
        "orderId": {
          "type": "string",
          "title": "uuid"
        },
        "userId": {
          "type": "string",
          "title": "uuid"
        },
        "amount": {
          "type": "number",
          "format": "double"
        },
        "status": {
          "type": "string"
        },
        "payloadJson": {
          "type": "string",
          "title": "raw json as string"
        },
        "createdAt": {
          "type": "string",
          "title": "RFC3339"
        },
        "updatedAt": {
          "type": "string",
          "title": "RFC3339"
        },
        "bucket": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "v1UpsertOrderRequest": {
      "type": "object",
      "properties": {
        "orderId": {
          "type": "string",
          "title": "empty or \"0\" =\u003e create"
        },
        "userId": {
          "type": "string"
        },
        "status": {
          "type": "string",
          "title": "optional; \"deleted\" =\u003e soft-delete"
        },
        "payloadJson": {
          "type": "string",
          "title": "required"
        }
      }
    },
    "v1UpsertOrderResponse": {
      "type": "object",
      "properties": {
        "orderId": {
          "type": "string"
        },
        "status": {
          "type": "string",
          "title": "created|updated|deleted"
        }
      }
    }
  }
}